Handle mode == Auto
At least the CT50 that I have only returns the corresponding setpoint
when in either heating or cooling mode. Get both modes' setpoints on
startup.

When setting a setpoint, it will switch to the corresponding mode so if
we were in auto initially, go back to it after setting that setpoint